+++ test/app/components/plugin/env-switcher.test.js
@@ -319,5 +319,33 @@ describe('Environment Switcher', () => {
       const publishButton = recursiveQuery(notificationItem, 'sp-action-button');
       expect(publishButton.hasAttribute('disabled')).to.be.true;
     }).timeout(20000);
+
+    it('should not show edit if byom sourceLocation', async () => {
+      sidekickTest
+        .mockFetchSidekickConfigSuccess()
+        .mockFetchStatusSuccess(false, {
+          webPath: '/query-index.json',
+          resourcePath: '/query-index.json',
+          preview: {
+            url: 'https://main--blog--adobecom.hlx.page/query-index.json',
+            status: 200,
+            contentBusId: 'helix-content-bus/cbid/preview/query-index.json',
+            contentType: 'application/json',
+            sourceLocation: 'markup:https://byom.adobeioruntime.net/api/v1/web/convert/main/index.html?wcmmode=disabled',
+            lastModified: 'Tue, 12 Sep 2023 19:38:51 GMT',
+            permissions: [
+              'read',
+              'write',
+            ],
+          },
+        }, HelixMockContentSources.SHAREPOINT, 'https://admin.hlx.page/status/adobe/aem-boilerplate/main/placeholders.json')
+        .mockHelixEnvironment(HelixMockEnvironments.PREVIEW, HelixMockContentType.SHEET);
+
+      sidekick = sidekickTest.createSidekick();
+      await sidekickTest.awaitEnvSwitcher();
+
+      const editPlugin = recursiveQuery(sidekick, '.env-edit');
+      expect(editPlugin).to.be.undefined;
+    });
   });
 });